Move Stripe to compatible package
Stripe uses Node’s crypto which clashes with React Native. Also fix phantom dependencies.
