+++ sdk/package.json
@@ -46,18 +46,26 @@
     "@hike/utils": "workspace:*",
     "@react-email/components": "0.0.25",
     "@react-email/render": "1.0.1",
+    "@tabler/icons-react": "^2",
+    "react-email": "^3.0.1",
     "jose": "^5.2.4",
+    "dayjs": "^1.11.11",
     "stripe": "^16.2.0"
   },
   "peerDependencies": {
     "@mantine/core": "^7.13.2",
-    "@mantine/hooks": "^7.13.2"
+    "@mantine/hooks": "^7.13.2",
+    "react": "^18",
+    "react-dom": "^18"
   },
   "devDependencies": {
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
     "@jest/globals": "^29.7.0",
     "@types/jest": "^29.5.12",
     "jest": "^29.7.0",
-    "react-email": "^3.0.1",
+    "react": "^18",
+    "react-dom": "^18",
     "ts-jest": "^29.1.2"
   },
   "jest": {
