+++ sdk/src/index.ts
@@ -36,4 +36,11 @@ export * from '@hike/utils';
 // Export platform specific functionality
 export * from './auth/extractToken';
 export * from './auth/verifyToken';
+export * from './billing/hooks/useGetBillingOverview';
+export * from './billing/hooks/useGetStripeEntities';
+export * from './billing/hooks/useGetStripeInvoice';
+export * from './billing/hooks/useGetSubscriptionInvoiceStats';
+export * from './billing/hooks/useGetUpFrontPaymentInfo';
+export * from './billing/hooks/useInvoices';
 export * from './billing/stripe-service';
+
